pyproject.toml
[tool.poetry]
name = "llm-snowglobe"
version = "1.0.2-dev"
description = "Snow Globe multi-agent system for open-ended wargames with large language models"
license = "Apache-2.0"
authors = ["Daniel Hogan <6313241+dphogan@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/IQTLabs/snowglobe"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
numpy = "*"
torch = "*"
torchvision = "*"
transformers = "*"
accelerate = "*"
tqdm = "*"
pyyaml = "*"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
platformdirs = "*"
langchain = "0.3.27"
openai = "*"
twine = "*"
jupyter = "*"
nicegui = "2.24.2"
markdown2 = "*"
watchfiles = "*"
langgraph = "*"
unstructured = "*"
langchain-openai = "*"
langchain-google-genai = "<3"
langchain-huggingface = "*"
langchain-chroma = "*"
langchain-community = "*"
langchain-text-splitters = "*"
llama-cpp-python = ">=0.3.8"
pdbpp = "*"
ruamel-yaml = "*"
black = "^25.9.0"
ruff = "^0.13.2"

[tool.poetry.scripts]
snowglobe_config = "llm_snowglobe.scripts:config"
snowglobe_server = "llm_snowglobe.scripts:server"
snowglobe_game = "llm_snowglobe.user_defined_game:main"
snowglobe_simulation = "llm_snowglobe.user_defined_sim:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
